WIP refactor for cleaning code
