+++ koin-projects/examples/coffee-maker/src/main/kotlin/org/koin/example/CoffeeMaker.kt
@@ -2,7 +2,9 @@ package org.koin.example
 
 import org.koin.ksp.KoinInject
 
-@KoinInject
+@KoinInject(
+    isSingle = true, bind = Pump::class
+)
 class CoffeeMaker(private val pump: Pump, private val heater: Heater) {
 
     fun brew() {
