+++ koin-projects/examples/coffee-maker/src/main/kotlin/org/koin/example/Whatever.kt
@@ -1,0 +1,22 @@
+package org.koin.example
+
+import java.lang.ref.WeakReference
+
+/**
+ * @author Fabio de Matos
+ * @since 21/04/2021.
+ */
+class Whatever {
+
+    fun what(): String {
+        return "this"
+    }
+
+    fun <T> who() {
+
+    }
+
+    val propp = "ertiy"
+
+    lateinit var teest : WeakReference<List<Int >>
+}
