+++ koin-projects/koin-ksp/src/main/kotlin/org/koin/ksp/TestProcessor.kt
@@ -1,7 +1,6 @@
 package org.koin.ksp
 
 import com.google.devtools.ksp.closestClassDeclaration
-import com.google.devtools.ksp.getAllSuperTypes
 import com.google.devtools.ksp.getClassDeclarationByName
 import com.google.devtools.ksp.getConstructors
 import com.google.devtools.ksp.processing.*
@@ -35,7 +34,13 @@ open class TestProcessor : SymbolProcessor {
             codeGenerator.createNewFile(Dependencies(false), "", "TestProcessor_finish", "log")
         emit("Finish")
 
-        writeKoinModule()
+        KoinModuleWriter(
+            logFile
+        ).writeKoinModule(
+            codeGenerator,
+            rootDependencyList,
+            indirectDependencyList
+        )
 
         emit("Injector over ${activityInjectors.keys}", "++")
         InjectorWriter(codeGenerator).writeKoinInjectors(activityInjectors)
@@ -85,11 +90,14 @@ open class TestProcessor : SymbolProcessor {
 
         emit("TestProcessor: process", "")
 
+//        FindKoinModulesUseCase(logFile)
+//            .invoke(resolver)
+
         resolver
             .getSymbolsWithAnnotation(KoinInject::class.java.canonicalName)
             .map {
                 it.accept(
-                    KoinVisitor(it,
+                    KoinVisitor(
                         lambdaClass = { clazz ->
                             emit("Found something for target ${clazz.closestClassDeclaration()}")
 
@@ -204,154 +212,6 @@ open class TestProcessor : SymbolProcessor {
             }
 
         depthCount.decrementAndGet()
-
-    }
-
-    private fun writeKoinModule() {
-
-        emit("Writing module: 1- $rootDependencyList")
-        emit("Writing module: 2- $indirectDependencyList")
-
-        val fileKt = codeGenerator.createNewFile(
-            Dependencies(false),
-            "org.koin.ksp.generated",
-            "KoinKspModule",
-            "kt"
-        )
-
-        fileKt.appendText("package org.koin.ksp.generated\n")
-        fileKt.appendText("\n")
-        fileKt.appendText("\n")
-        fileKt.appendText("import org.koin.dsl.module\n")
-        fileKt.appendText("\n")
-        fileKt.appendText("\n")
-        fileKt.appendText("class KoinKspModule {\n")
-        fileKt.appendText("val module = module { \n", identation = "\t")
-
-        emit("Hello write koin")
-        createFactories(fileKt, rootDependencyList, indirectDependencyList)
-
-        fileKt.appendText("}\n", identation = "\t")
-        fileKt.appendText("}\n")
-        fileKt.close()
-
-        emit("Writing module done : $rootDependencyList")
-    }
-
-    private fun createFactories(
-        fileKt: OutputStream,
-        injectedFieldList: Set<KSPropertyDeclaration>,
-        injectedIndirectClasses: Set<KSClassDeclaration>
-    ) {
-
-        emit("Factories now $injectedFieldList")
-
-        injectedFieldList
-            .map {
-                emit("Factory for $it")
-
-                createFactory(fileKt, it)
-
-            }
-
-        injectedIndirectClasses
-            .map {
-                createFactory(fileKt, it)
-            }
-    }
-
-    /**
-     * Creates a singleton out of [injectedReference]
-     */
-    private fun createFactory(fileKt: OutputStream, injectedReference: KSClassDeclaration) {
-
-
-        val fullClassName = injectedReference
-            .closestClassDeclaration()
-            .let {
-                "${it?.qualifiedName?.getQualifier()}.$it"
-            }
-
-        val location = injectedReference.location
-
-        injectedReference
-            .closestClassDeclaration()
-            ?.primaryConstructor
-            ?.parameters
-            ?.size
-            ?.let {
-                Array(it) {
-                    "get()"
-                }
-            }
-            ?.joinToString(separator = ", ")
-            .let {
-
-                "single { $fullClassName($it) }"
-            }
-            .let {
-                fileKt.appendText("\n\n")
-                fileKt.appendText("// location $location\n", identation = "\t\t")
-                fileKt.appendText("$it\n", identation = "\t\t")
-            }
-
-
-        createSuperFactory(fileKt, injectedReference)
-    }
-
-    /**
-     * Creates a factory out of the supertypes of [injectedReference]
-     */
-    fun createSuperFactory(fileKt: OutputStream, injectedReference: KSClassDeclaration) {
-
-        injectedReference
-            .superTypes
-            .mapNotNull {
-                emit("Super types for $injectedReference = $it", "\t")
-                it.resolve().declaration.closestClassDeclaration()
-            }
-            .map {
-
-                emit("Factory with $injectedReference", "\t\t")
-                "factory<${it.toFullName()}> { get<${injectedReference.toFullName()}>() } "
-            }
-            .map {
-                fileKt.appendText("$it\n", identation = "\t\t")
-            }
-    }
-
-    private fun createFactory(fileKt: OutputStream, injectedReference: KSPropertyDeclaration) {
-
-        val clazz = injectedReference
-            .type
-            .resolve()
-            .declaration
-            .qualifiedName
-            ?.let { "${it.getQualifier()}.${it.getShortName()}" }
-
-
-        val location = injectedReference.location
-
-
-        injectedReference.javaClass
-            .constructors
-            .first()
-            .parameterCount
-            .let {
-                Array(it) {
-                    "get()"
-                }
-            }
-            .joinToString(separator = ", ")
-            .let {
-
-                "single(){ $clazz($it) }"
-            }
-            .let {
-                fileKt.appendText("\n")
-                fileKt.appendText("// llocation $location\n", identation = "\t\t")
-                fileKt.appendText("$it\n", identation = "\t\t")
-            }
 
     }
